+++ fiddles/angular2/fiddle-0038-DragAndDrop/app/util/util.ts
@@ -1,0 +1,19 @@
+
+export class Util {
+
+    static get radianToDegreeConstant():number {
+        return 180 / Math.PI;
+    }
+
+    static constrain(num:number, min:number, max:number):number {
+        let x = parseFloat(num);
+        if (min === null) {
+            min = num;
+        }
+        if (max === null) {
+            max = num;
+        }
+        return (x < min) ? min : ((x > max) ? max : x);
+    }
+
+}