+++ fiddles/node/fiddle-0005-ReadStream/app.js
@@ -1,0 +1,14 @@
+#!/usr/bin/env node --harmony
+
+"use strict";
+
+const fs = require('fs'),
+    stream = fs.createReadStream(process.argv[2]);
+
+stream.on('data', function(chunk) {
+    process.stdout.write(chunk);
+});
+
+stream.on('error', function(err) {
+    process.stderr.write("ERROR:  " + err.message + "\n");
+});